Update pcgi utility: can't kill itself for -f parameters
git-svn-id: https://psinterop.svn.codeplex.com/svn@17518 a83551a4-30f6-4d81-a974-c6ced450ddbf

+++ pcgi/Program.cs
@@ -185,7 +185,8 @@ Goodbye, World!
                                     continue;
                             }
 
-                            if (kill)
+                            int currentId = System.Diagnostics.Process.GetCurrentProcess().Id;
+                            if (kill && p.ProcessId != currentId)
                                 Win32Wrapper.TerminateProcess(p);
 
                             string message = string.Format("{0}{1} ({2}) {3}",
